pdf: Add rudiments for coordinate space handling
Performance is not as huge of a concern at this point, but I do want the
core architecture to be able to parse PDF files as fast as possible.
Slices/fixed-size arrays seem like the logical way to do that.  I'm sure
that there will be optimizations to consider, but for now this is how
I'll implement this.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>